# docs/openapi.json
{
  "openapi" : "3.1.0",
...
  "components" : {
    "schemas" : {
      "AttendanceStatus" : {
        "type" : "string",
        "enum" : [ "PRESENT", "ABSENT", "EXCUSED" ]
      },
...
      "CreateTeamMembershipRequest" : {
        "type" : "object",
        "required" : [ "teamId", "personId", "role" ],
        "properties" : {
          "teamId" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "personId" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "role" : {
            "$ref" : "#/components/schemas/TeamMemberRole"
          },
          "joinedAt" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "leftAt" : {
            "$ref" : "#/components/schemas/LocalDate"
          }
        }
      },
...
      "CreateTrainingParticipantRequest" : {
        "type" : "object",
        "required" : [ "trainingId", "personId", "attendanceStatus" ],
        "properties" : {
          "trainingId" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "personId" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "attendanceStatus" : {
            "$ref" : "#/components/schemas/AttendanceStatus"
          },
          "notes" : {
            "type" : "string"
          }
        }
      },
...
      "DeactivateTeamMembershipRequest" : {
        "type" : "object",
        "properties" : {
          "leftAt" : {
            "$ref" : "#/components/schemas/LocalDate"
          }
        }
      },
      "ErrorResponse" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "error" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "timestamp" : {
            "$ref" : "#/components/schemas/Instant"
          }
        }
      },
      "Instant" : {
        "type" : "string",
        "format" : "date-time",
        "examples" : [ "2022-03-10T16:15:50Z" ]
      },
      "LocalDate" : {
        "type" : "string",
        "format" : "date",
        "examples" : [ "2022-03-10" ]
      },
      "LocalTime" : {
        "type" : "string",
        "format" : "local-time",
        "examples" : [ "13:45:30.123456789" ],
        "externalDocs" : {
          "description" : "As defined by 'partial-time' in RFC3339",
          "url" : "https://www.rfc-editor.org/rfc/rfc3339.html#section-5.6"
        }
      },
      "PersonDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "firstName" : {
            "type" : "string"
          },
          "lastName" : {
            "type" : "string"
          },
          "birthDate" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "nickname" : {
            "type" : "string"
          },
          "notes" : {
            "type" : "string"
          },
          "phoneNumbers" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PhoneNumberDto"
            }
          },
          "createdAt" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "updatedAt" : {
            "$ref" : "#/components/schemas/Instant"
          }
        }
      },
...
      "PhoneNumberDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "type" : {
            "$ref" : "#/components/schemas/PhoneType"
          },
          "number" : {
            "type" : "string"
          }
        }
      },
      "PhoneType" : {
        "type" : "string",
        "enum" : [ "MOBILE", "HOME", "WORK", "EMERGENCY" ]
      },
...
      "TeamDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "seasonId" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "seasonName" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "createdAt" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "updatedAt" : {
            "$ref" : "#/components/schemas/Instant"
          }
        }
      },
      "TeamMemberRole" : {
        "type" : "string",
        "enum" : [ "PLAYER", "COACH", "ASSISTANT_COACH" ]
      },
      "TeamMembershipDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "team" : {
            "$ref" : "#/components/schemas/TeamDto"
          },
          "person" : {
            "$ref" : "#/components/schemas/PersonDto"
          },
          "role" : {
            "$ref" : "#/components/schemas/TeamMemberRole"
          },
          "joinedAt" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "leftAt" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "createdAt" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "updatedAt" : {
            "$ref" : "#/components/schemas/Instant"
          }
        }
      },
      "TrainingDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "team" : {
            "$ref" : "#/components/schemas/TeamDto"
          },
          "trainingDate" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "startTime" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "endTime" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "location" : {
            "type" : "string"
          },
          "notes" : {
            "type" : "string"
          },
          "status" : {
            "$ref" : "#/components/schemas/TrainingStatus"
          },
          "createdAt" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "updatedAt" : {
            "$ref" : "#/components/schemas/Instant"
          }
        }
      },
      "TrainingParticipantDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "training" : {
            "$ref" : "#/components/schemas/TrainingDto"
          },
          "person" : {
            "$ref" : "#/components/schemas/PersonDto"
          },
          "attendanceStatus" : {
            "$ref" : "#/components/schemas/AttendanceStatus"
          },
          "notes" : {
            "type" : "string"
          },
          "createdAt" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "updatedAt" : {
            "$ref" : "#/components/schemas/Instant"
          }
        }
      },
      "TrainingStatus" : {
        "type" : "string",
        "enum" : [ "COMPLETED", "CANCELLED" ]
      },
      "UUID" : {
        "type" : "string",
        "format" : "uuid",
        "pattern" : "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
      },
...
    }
  },
  "paths" : {
...
    "/api/v1/team-memberships" : {
      "get" : {
        "summary" : "Find team memberships",
        "description" : "Returns team memberships matching the optional search criteria.",
        "tags" : [ "Team Memberships" ],
        "parameters" : [ {
          "name" : "active",
          "in" : "query",
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "name" : "personId",
          "in" : "query",
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        }, {
          "name" : "role",
          "in" : "query",
          "schema" : {
            "$ref" : "#/components/schemas/TeamMemberRole"
          }
        }, {
          "name" : "teamId",
          "in" : "query",
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Team memberships found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/TeamMembershipDto"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "summary" : "Create team membership",
        "description" : "Adds a person to a team with a role.",
        "tags" : [ "Team Memberships" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateTeamMembershipRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Team membership created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TeamMembershipDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Referenced person or team not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "Team membership already exists",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/team-memberships/{id}/deactivate" : {
      "patch" : {
        "summary" : "Deactivate team membership",
        "description" : "Sets the left date of a team membership.",
        "tags" : [ "Team Memberships" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DeactivateTeamMembershipRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Team membership deactivated",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TeamMembershipDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid left date",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Team membership not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/training-participants" : {
      "get" : {
        "summary" : "Find training participants",
        "description" : "Returns training participants matching the optional search criteria.",
        "tags" : [ "Training Participants" ],
        "parameters" : [ {
          "name" : "attendanceStatus",
          "in" : "query",
          "schema" : {
            "$ref" : "#/components/schemas/AttendanceStatus"
          }
        }, {
          "name" : "personId",
          "in" : "query",
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        }, {
          "name" : "trainingId",
          "in" : "query",
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Training participants found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/TrainingParticipantDto"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "summary" : "Create training participant",
        "description" : "Adds a person to a training with an attendance status.",
        "tags" : [ "Training Participants" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateTrainingParticipantRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Training participant created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TrainingParticipantDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Referenced training or person not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "Training participant already exists",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/training-participants/{id}" : {
      "delete" : {
        "summary" : "Delete training participant",
        "description" : "Deletes a mistakenly created training participant entry.",
        "tags" : [ "Training Participants" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Training participant deleted"
          },
          "404" : {
            "description" : "Training participant not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/trainings/{id}" : {
      "get" : {
        "summary" : "Find training by id",
        "description" : "Returns a single training by id.",
        "tags" : [ "Trainings" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Training found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TrainingDto"
                }
              }
            }
          },
          "404" : {
            "description" : "Training not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "summary" : "Delete training",
        "description" : "Deletes a mistakenly created training including its participants.",
        "tags" : [ "Trainings" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Training deleted"
          },
          "404" : {
            "description" : "Training not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "info" : {
    "title" : "Coach Buddy API",
    "version" : "1.0.0",
    "description" : "API für Trainings-, Team- und Personenverwaltung"
  }
}